requirements.txt
Babel==2.7.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-multiselectfield==0.1.9
django-phonenumber-field==3.0.1
gunicorn==19.9.0
idna==2.8
phonenumbers==8.10.20
psycopg2-binary==2.8.3
python-decouple==3.1
pytz==2019.3
requests==2.22.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
